move types and operations to dedicated files for readbility
